for screen restrcit to stage and file strcuture for screeen and many more WIP

+++ src/extension.ts
@@ -16,6 +16,7 @@ const decoder = new util.TextDecoder('utf-8');
 import { DeviceFs, MonocleFile,ScreenProvider,DeviceInfoProvider } from './fileSystemProvider';
 export const monocleFolder = "device files";
 export const monocleFiles = '.brilliant/device-files.json';
+export const monocleScreens = '.brilliant/screens.json';
 export const screenFolder ="screens";
 let statusBarItemBle:vscode.StatusBarItem;
 
@@ -40,18 +41,57 @@ function isEmpty(obj:object) {
   
 	return true;
   }
+
+  export const configScreenReadUpdate = async  function(dataToupdate?:object):Promise<object>{
+	let workspaces = vscode.workspace.workspaceFolders;
+	let newdata:any = {};
+	if(workspaces){
+		
+		let _configFileScreen = vscode.Uri.joinPath(workspaces[0].uri,monocleScreens);
+		if(dataToupdate && !(await isPathExist(_configFileScreen)) && !isEmpty(dataToupdate)){
+			await vscode.workspace.fs.writeFile(_configFileScreen,Buffer.from(JSON.stringify({}, null, "\t")));
+		}
+	
+		
+		try{ 
+			if((await isPathExist(_configFileScreen))){
+				let data =  await vscode.workspace.fs.readFile(_configFileScreen);
+				let jsonData = JSON.parse(decoder.decode(data));
+				newdata = {...jsonData};
+			}
+
+		} catch (error) {
+			
+		}
+		
+		if(dataToupdate && (await isPathExist(_configFileScreen))){
+			newdata  = {...newdata,...dataToupdate};
+			Object.keys(newdata).forEach(function(skey){
+				if(newdata[skey]===false){
+					delete newdata[skey];
+				}
+			});
+			await vscode.workspace.fs.writeFile(_configFileScreen,Buffer.from(JSON.stringify(newdata, null, "\t")));
+		}
+		return newdata;
+		
+	}
+	return newdata;
+	
+};
 export const configReadUpdate = async  function(dataToupdate?:object):Promise<object>{
 	let workspaces = vscode.workspace.workspaceFolders;
 	let newdata:any = {};
 	if(workspaces){
 		
 		let _configFile = vscode.Uri.joinPath(workspaces[0].uri,monocleFiles);
+		let _configFileScreen = vscode.Uri.joinPath(workspaces[0].uri,monocleScreens);
 		if(dataToupdate && !(await isPathExist(_configFile)) && !isEmpty(dataToupdate)){
 			await vscode.workspace.fs.writeFile(_configFile,Buffer.from(JSON.stringify({}, null, "\t")));
 		}
 	
 		
-		try {
+		try{ 
 			if((await isPathExist(_configFile))){
 				let data =  await vscode.workspace.fs.readFile(_configFile);
 				let jsonData = JSON.parse(decoder.decode(data));
@@ -457,6 +497,7 @@ export async function activate(context: vscode.ExtensionContext) {
 		vscode.workspace.onWillDeleteFiles(async (e:vscode.FileDeleteEvent)=>{
 			if(vscode.workspace.workspaceFolders){
 				let newpathKey:any = await configReadUpdate();
+				let newpathScreens:any = await configReadUpdate();
 				let projectPath = vscode.workspace.workspaceFolders[0].uri;
 				for (let index = 0; index < e.files.length; index++) {
 					const ef = e.files[index];
@@ -470,12 +511,18 @@ export async function activate(context: vscode.ExtensionContext) {
 						if(Object.keys(newpathKey).includes(localDir)){
 							newpathKey[localDir] =false;
 						}
+						let basename = path.posix.basename(localDir);
+						let segments = basename.split("/");
+						if(segments.length>0 && Object.keys(newpathScreens).includes(segments[segments.length-1].replace('.py',''))){
+							newpathScreens[segments[segments.length-1].replace('.py','')] = false;
+						}
 					}
 							// }
 						// await memFs.deleteFile(devicePath);
 					// }
 				}
 				await configReadUpdate(newpathKey);
+				await configReadUpdate(newpathScreens);
 			}
 			
 		}),
@@ -853,16 +900,31 @@ export async function activate(context: vscode.ExtensionContext) {
 		// for UI webview
 		vscode.commands.registerCommand("brilliant-ar-studio.openUIEditor", async () => {
 			if(vscode.workspace.workspaceFolders){
-				let screenName = await vscode.window.showInputBox({prompt:"Enter Screen name"});
+				let screenName = await vscode.window.showInputBox({prompt:"Enter Screen name",title:"New Screen"});
+				const projectPath = vscode.workspace.workspaceFolders[0].uri;
 				if(screenName){
 					screenName = screenName.replaceAll(" ","_").replaceAll("/","").replaceAll("\\","").replaceAll("-","_");
-					let screenPath = vscode.Uri.joinPath(vscode.workspace.workspaceFolders[0].uri,screenFolder,screenName+"_screen.py");
-					await vscode.workspace.fs.writeFile(screenPath,Buffer.from('# GENERATED BRILLIANT AR STUDIO Do not modify this file directly\n\nimport display\n\nclass '+screenName+':\n\tpass'));
-					await vscode.commands.executeCommand('vscode.open',screenPath,vscode.ViewColumn.One);
-					UIEditorPanel.render(context.extensionUri,screenName,screenPath);
-					screenProvider.refresh();
+					let filename = screenName.endsWith('.py')?screenName:screenName+".py";
+					let screenPath = vscode.Uri.joinPath(projectPath,filename);
 					
+					if(!await isPathExist(screenPath)){
+						let localPath = screenPath.fsPath.replace(projectPath.fsPath, "").replaceAll("\\","/");
+						let screenKey:any = {};
+						let pathKey:any = {};
+						screenKey[screenName] = {"filePath":screenPath.path};
+						pathKey[localPath] = filename;
+						await vscode.workspace.fs.writeFile(screenPath,Buffer.from('# GENERATED BRILLIANT AR STUDIO Do not modify this file directly\n\nimport display\n\nclass '+screenName+':\n\tpass'));
+						await configScreenReadUpdate(screenKey);
+						await configReadUpdate(pathKey);
+						await vscode.commands.executeCommand('vscode.open',screenPath,vscode.ViewColumn.One);
+						UIEditorPanel.render(context.extensionUri,screenName,screenPath);
+						screenProvider.refresh();
+					}else{
+						vscode.window.showErrorMessage(filename+' already exist');
+					}
 				}
+			}else{
+				vscode.window.showWarningMessage("First initialize a project");
 			}
 			
 		}),
